Update visualize_hifisomatic.qmd
Had a case where the "PS" column was NA rather than TRUE or FALSE after the `all(x == ".")` function. Having an NA rather than a T/F value caused subsequent steps to fail.

This occurred because the filtered table_to_display had a PS column with only NA values. Thus, when filtering for names that only equals TRUE...this column was retained, but the column name was not properly pulled leading to a value of NA in the "col_wit_onlyNA" vector...which is not a valid column name for the next filtering step below. 

# Select only relevant cols
table_to_disp <- table_to_disp %>%
    select(
        cols_to_highlight,
        rest_of_cols
    ) %>%
    select(!all_of(col_with_onlyNA))